feat: Links screen is set up to contain links
Modified the links fragment into a RelativeLayout so that the ad will always be displayed at the bottom of the screen. Added a child LinearLayout with a vertical orientation for easy adding of more links in the future. Added one hyperlink which needed to be done programatically due to xml limitations with hyperlinks.